refactor(adapter-openclaw): extract image inference into image-inference.ts
Fifth cut of the runtime.ts god-file split. Moves the 13 image
provider/model arg formatters + output-path setup + CLI-output parsers
(defaultOpenClawImageOutputPath, normalizeOpenClawOutputFormat,
openClawImageModelArg, providerFromImageModel, modelNameFromImageModel,
parseOpenClawImageProviders, imageQualityFromMetadata, tagRuntimeServed,
parseOpenClawImageResult, collectOpenClawImageFiles, openClawImageFileCandidate,
imageMimeTypeForPath, openClawImageProviderText) into a new image-inference.ts.

The class's image methods own the exec + provider TTL cache and import the
8 helpers they call back; the recursive file-collection internals stay
module-private. runtime.ts shed its now-unused mkdtempSync/tmpdir/firstString
imports.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 2,393 -> 2,241.
Verified: typecheck, eslint (changed files), adapter suite 115-0 (--isolate),
full suite 5124-0, binary build (stamps reverted), boot smoke (adapter loads,
schedule activates, GET /api/plugins/schedule/ -> 200).
